+++ src/android/AppboyPlugin.java
@@ -5,33 +5,31 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Build;
-import android.net.Uri;
 import android.util.Log;
 
-import com.appboy.Appboy;
-import com.appboy.AppboyUser;
-import com.appboy.configuration.AppboyConfig;
 import com.appboy.enums.CardCategory;
 import com.appboy.enums.Gender;
 import com.appboy.enums.Month;
 import com.appboy.enums.NotificationSubscriptionType;
 import com.appboy.enums.SdkFlavor;
-import com.appboy.events.ContentCardsUpdatedEvent;
 import com.appboy.events.FeedUpdatedEvent;
 import com.appboy.events.IEventSubscriber;
 import com.appboy.models.cards.Card;
-import com.appboy.models.outgoing.AppboyProperties;
 import com.appboy.models.outgoing.AttributionData;
-import com.appboy.services.AppboyLocationService;
-import com.appboy.support.AppboyLogger;
-import com.appboy.ui.activities.AppboyContentCardsActivity;
 import com.appboy.ui.activities.AppboyFeedActivity;
-import com.appboy.ui.inappmessage.AppboyInAppMessageManager;
+import com.braze.Braze;
+import com.braze.BrazeUser;
+import com.braze.configuration.BrazeConfig;
+import com.braze.enums.BrazeSdkMetadata;
+import com.braze.events.ContentCardsUpdatedEvent;
+import com.braze.models.outgoing.BrazeProperties;
+import com.braze.support.BrazeLogger;
+import com.braze.ui.activities.ContentCardsActivity;
+import com.braze.ui.inappmessage.BrazeInAppMessageManager;
 
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.CordovaPreferences;
-import org.apache.cordova.PluginResult;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -43,7 +41,7 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 public class AppboyPlugin extends CordovaPlugin {
-  private static final String TAG = String.format("Appboy.%s", AppboyPlugin.class.getName());
+  private static final String TAG = "BrazeCordova";
 
   // Runtime permissions
   private static final int LOCATION_REQUEST_CODE = 2;
@@ -61,6 +59,12 @@ public class AppboyPlugin extends CordovaPlugin {
   private static final String CUSTOM_API_ENDPOINT_PREFERENCE = "com.appboy.android_api_endpoint";
   private static final String ENABLE_LOCATION_PREFERENCE = "com.appboy.enable_location_collection";
   private static final String ENABLE_GEOFENCES_PREFERENCE = "com.appboy.geofences_enabled";
+  private static final String DISABLE_AUTO_START_SESSIONS_PREFERENCE = "com.appboy.android_disable_auto_session_tracking";
+  /**
+   * When applied, restricts the SDK from taking
+   * focus away from the Cordova WebView on affected API versions.
+   */
+  private static final String ENABLE_CORDOVA_WEBVIEW_REQUEST_FOCUS_FIX_PREFERENCE = "com.braze.android_apply_cordova_webview_focus_request_fix";
 
   // Numeric preference prefix
   private static final String NUMERIC_PREFERENCE_PREFIX = "str_";
@@ -79,8 +83,9 @@ public class AppboyPlugin extends CordovaPlugin {
   private static final String LOG_CONTENT_CARDS_DISMISSED_METHOD = "logContentCardDismissed";
 
   private boolean mPluginInitializationFinished = false;
+  private boolean mDisableAutoStartSessions = false;
   private Context mApplicationContext;
-  private Map<String, IEventSubscriber<FeedUpdatedEvent>> mFeedSubscriberMap = new ConcurrentHashMap<>();
+  private final Map<String, IEventSubscriber<FeedUpdatedEvent>> mFeedSubscriberMap = new ConcurrentHashMap<>();
 
   private CallbackContext locationPermissionCallbackContext;
   private CallbackContext geofenceCallbackContext;
@@ -90,13 +95,12 @@ public class AppboyPlugin extends CordovaPlugin {
     mApplicationContext = this.cordova.getActivity().getApplicationContext();
 
     // Configure Appboy using the preferences from the config.xml file passed to our plugin
-    configureAppboyFromCordovaPreferences(this.preferences);
+    configureFromCordovaPreferences(this.preferences);
 
     // Since we've likely passed the first Application.onCreate() (due to the plugin lifecycle), lets call the
     // in-app message manager and session handling now
-    AppboyInAppMessageManager.getInstance().registerInAppMessageManager(this.cordova.getActivity());
-
     initializeGeofences();
+    BrazeInAppMessageManager.getInstance().registerInAppMessageManager(this.cordova.getActivity());
     mPluginInitializationFinished = true;
   }
 
@@ -126,20 +130,22 @@ public class AppboyPlugin extends CordovaPlugin {
     initializePluginIfAppropriate();
     Log.i(TAG, "Received " + action + " with the following arguments: " + args);
 
-    // Appboy methods
     switch (action) {
+      case "startSessionTracking":
+        mDisableAutoStartSessions = false;
+        return true;
       case "registerAppboyPushMessages":
-        Appboy.getInstance(mApplicationContext).registerAppboyPushMessages(args.getString(0));
+        Braze.getInstance(mApplicationContext).registerAppboyPushMessages(args.getString(0));
         return true;
       case "changeUser":
-        Appboy.getInstance(mApplicationContext).changeUser(args.getString(0));
+        Braze.getInstance(mApplicationContext).changeUser(args.getString(0));
         return true;
       case "logCustomEvent": {
-        AppboyProperties properties = null;
+        BrazeProperties properties = null;
         if (args.get(1) != JSONObject.NULL) {
-          properties = new AppboyProperties(args.getJSONObject(1));
+          properties = new BrazeProperties(args.getJSONObject(1));
         }
-        Appboy.getInstance(mApplicationContext).logCustomEvent(args.getString(0), properties);
+        Braze.getInstance(mApplicationContext).logCustomEvent(args.getString(0), properties);
         return true;
       }
       case "logPurchase": {
@@ -151,32 +157,32 @@ public class AppboyPlugin extends CordovaPlugin {
         if (args.get(3) != JSONObject.NULL) {
           quantity = args.getInt(3);
         }
-        AppboyProperties properties = null;
+        BrazeProperties properties = null;
         if (args.get(4) != JSONObject.NULL) {
-          properties = new AppboyProperties(args.getJSONObject(4));
+          properties = new BrazeProperties(args.getJSONObject(4));
         }
-        Appboy.getInstance(mApplicationContext).logPurchase(args.getString(0), currencyCode, new BigDecimal(args.getDouble(1)), quantity, properties);
+        Braze.getInstance(mApplicationContext).logPurchase(args.getString(0), currencyCode, new BigDecimal(args.getDouble(1)), quantity, properties);
         return true;
       }
       case "wipeData":
-        Appboy.wipeData(mApplicationContext);
+        Braze.wipeData(mApplicationContext);
         mPluginInitializationFinished = false;
         return true;
       case "enableSdk":
-        Appboy.enableSdk(mApplicationContext);
+        Braze.enableSdk(mApplicationContext);
         return true;
       case "disableSdk":
-        Appboy.disableSdk(mApplicationContext);
+        Braze.disableSdk(mApplicationContext);
         mPluginInitializationFinished = false;
         return true;
       case "requestImmediateDataFlush":
-        Appboy.getInstance(mApplicationContext).requestImmediateDataFlush();
+        Braze.getInstance(mApplicationContext).requestImmediateDataFlush();
         return true;
       case "requestContentCardsRefresh":
-        Appboy.getInstance(mApplicationContext).requestContentCardsRefresh(false);
+        Braze.getInstance(mApplicationContext).requestContentCardsRefresh(false);
         return true;
       case "getDeviceId":
-        callbackContext.success(Appboy.getInstance(mApplicationContext).getDeviceId());
+        callbackContext.success(Braze.getInstance(mApplicationContext).getDeviceId());
         return true;
       case "requestLocationPermission":
         requestLocationPermission(callbackContext);
@@ -190,8 +196,8 @@ public class AppboyPlugin extends CordovaPlugin {
         return false;
     }
 
-    // Appboy User methods
-    AppboyUser currentUser = Appboy.getInstance(mApplicationContext).getCurrentUser();
+    // User methods
+    BrazeUser currentUser = Braze.getInstance(mApplicationContext).getCurrentUser();
     if (currentUser != null) {
       switch (action) {
         case "setUserAttributionData":
@@ -239,17 +245,32 @@ public class AppboyPlugin extends CordovaPlugin {
           return true;
         case "setGender":
           String gender = args.getString(0).toLowerCase();
-          if (gender.equals("m")) {
-            currentUser.setGender(Gender.MALE);
-          } else if (gender.equals("f")) {
-            currentUser.setGender(Gender.FEMALE);
+          switch (gender) {
+            case "f":
+              currentUser.setGender(Gender.FEMALE);
+              break;
+            case "m":
+              currentUser.setGender(Gender.MALE);
+              break;
+            case "n":
+              currentUser.setGender(Gender.NOT_APPLICABLE);
+              break;
+            case "o":
+              currentUser.setGender(Gender.OTHER);
+              break;
+            case "p":
+              currentUser.setGender(Gender.PREFER_NOT_TO_SAY);
+              break;
+            case "u":
+              currentUser.setGender(Gender.UNKNOWN);
+              break;
           }
           return true;
         case "addAlias":
           currentUser.addAlias(args.getString(0), args.getString(1));
           return true;
         case "setDateOfBirth":
-          Month month = parseMonth(args.getInt(1));
+          Month month = Month.getMonth(args.getInt(1) - 1);
           currentUser.setDateOfBirth(args.getInt(0), month, args.getInt(2));
           return true;
         case "setCountry":
@@ -297,6 +318,12 @@ public class AppboyPlugin extends CordovaPlugin {
         case "setLanguage":
           currentUser.setLanguage(args.getString(0));
           return true;
+        case "addToSubscriptionGroup":
+          currentUser.addToSubscriptionGroup(args.getString(0));
+          return true;
+        case "removeFromSubscriptionGroup":
+          currentUser.removeFromSubscriptionGroup(args.getString(0));
+          return true;
       }
     }
 
@@ -308,7 +335,7 @@ public class AppboyPlugin extends CordovaPlugin {
         this.cordova.getActivity().startActivity(intent);
         return true;
       case "launchContentCards":
-        intent = new Intent(mApplicationContext, AppboyContentCardsActivity.class);
+        intent = new Intent(mApplicationContext, ContentCardsActivity.class);
         this.cordova.getActivity().startActivity(intent);
         return true;
     }
@@ -327,7 +354,7 @@ public class AppboyPlugin extends CordovaPlugin {
       case GET_CONTENT_CARDS_FROM_CACHE_METHOD:
         return handleContentCardsUpdateGetters(action, callbackContext);
       case LOG_CONTENT_CARDS_DISPLAYED_METHOD:
-        Appboy.getInstance(mApplicationContext).logContentCardsDisplayed();
+        Braze.getInstance(mApplicationContext).logContentCardsDisplayed();
         return true;
       case LOG_CONTENT_CARDS_CLICKED_METHOD:
       case LOG_CONTENT_CARDS_DISMISSED_METHOD:
@@ -343,30 +370,34 @@ public class AppboyPlugin extends CordovaPlugin {
   public void onPause(boolean multitasking) {
     super.onPause(multitasking);
     initializePluginIfAppropriate();
-    AppboyInAppMessageManager.getInstance().unregisterInAppMessageManager(this.cordova.getActivity());
+    BrazeInAppMessageManager.getInstance().unregisterInAppMessageManager(this.cordova.getActivity());
   }
 
   @Override
   public void onResume(boolean multitasking) {
     super.onResume(multitasking);
     initializePluginIfAppropriate();
-    // Registers the AppboyInAppMessageManager for the current Activity. This Activity will now listen for
-    // in-app messages from Appboy.
-    AppboyInAppMessageManager.getInstance().registerInAppMessageManager(this.cordova.getActivity());
+    // Registers the BrazeInAppMessageManager for the current Activity. This Activity will now listen for
+    // in-app messages from Braze.
+    BrazeInAppMessageManager.getInstance().registerInAppMessageManager(this.cordova.getActivity());
   }
 
   @Override
   public void onStart() {
     super.onStart();
     initializePluginIfAppropriate();
-    Appboy.getInstance(mApplicationContext).openSession(this.cordova.getActivity());
+    if (!mDisableAutoStartSessions) {
+      Braze.getInstance(mApplicationContext).openSession(this.cordova.getActivity());
+    }
   }
 
   @Override
   public void onStop() {
     super.onStop();
     initializePluginIfAppropriate();
-    Appboy.getInstance(mApplicationContext).closeSession(this.cordova.getActivity());
+    if (!mDisableAutoStartSessions) {
+      Braze.getInstance(mApplicationContext).closeSession(this.cordova.getActivity());
+    }
   }
 
   /**
@@ -379,34 +410,30 @@ public class AppboyPlugin extends CordovaPlugin {
   }
 
   /**
-   * Calls {@link Appboy#configure(Context, AppboyConfig)} using the values found from the {@link CordovaPreferences}.
+   * Calls {@link Braze#configure(Context, BrazeConfig)} using the values found from the {@link CordovaPreferences}.
    *
    * @param cordovaPreferences the preferences used to initialize this plugin
    */
-  private void configureAppboyFromCordovaPreferences(CordovaPreferences cordovaPreferences) {
-    AppboyLogger.d(TAG, "Setting Cordova preferences: " + cordovaPreferences.getAll());
+  private void configureFromCordovaPreferences(CordovaPreferences cordovaPreferences) {
+    BrazeLogger.d(TAG, "Setting Cordova preferences: " + cordovaPreferences.getAll());
 
     // Set the log level
     if (cordovaPreferences.contains(APPBOY_LOG_LEVEL_PREFERENCE)) {
-      AppboyLogger.setLogLevel(cordovaPreferences.getInteger(APPBOY_LOG_LEVEL_PREFERENCE, Log.INFO));
+      BrazeLogger.setLogLevel(cordovaPreferences.getInteger(APPBOY_LOG_LEVEL_PREFERENCE, Log.INFO));
     }
 
-    // Set the custom endpoint
-    if (cordovaPreferences.contains(CUSTOM_API_ENDPOINT_PREFERENCE)) {
-      final String customApiEndpoint = cordovaPreferences.getString(CUSTOM_API_ENDPOINT_PREFERENCE, "");
-      if (!customApiEndpoint.equals("")) {
-        Appboy.setAppboyEndpointProvider(appboyEndpoint ->
-            appboyEndpoint.buildUpon()
-            .authority(customApiEndpoint).build()
-        );
-      }
+    // Disable auto starting sessions
+    if (cordovaPreferences.getBoolean(DISABLE_AUTO_START_SESSIONS_PREFERENCE, false)) {
+      BrazeLogger.d(TAG, "Disabling session auto starts");
+      mDisableAutoStartSessions = true;
     }
 
     // Set the values used in the config builder
-    AppboyConfig.Builder configBuilder = new AppboyConfig.Builder();
+    BrazeConfig.Builder configBuilder = new BrazeConfig.Builder();
 
     // Set the flavor
-    configBuilder.setSdkFlavor(SdkFlavor.CORDOVA);
+    configBuilder.setSdkFlavor(SdkFlavor.CORDOVA)
+                 .setSdkMetadata(EnumSet.of(BrazeSdkMetadata.CORDOVA));
 
     if (cordovaPreferences.contains(APPBOY_API_KEY_PREFERENCE)) {
       configBuilder.setApiKey(cordovaPreferences.getString(APPBOY_API_KEY_PREFERENCE, null));
@@ -438,15 +465,27 @@ public class AppboyPlugin extends CordovaPlugin {
     if (cordovaPreferences.contains(ENABLE_GEOFENCES_PREFERENCE)) {
       configBuilder.setGeofencesEnabled(cordovaPreferences.getBoolean(ENABLE_GEOFENCES_PREFERENCE, false));
     }
+    if (cordovaPreferences.contains(CUSTOM_API_ENDPOINT_PREFERENCE)) {
+      final String customApiEndpoint = cordovaPreferences.getString(CUSTOM_API_ENDPOINT_PREFERENCE, "");
+      if (!customApiEndpoint.equals("")) {
+        configBuilder.setCustomEndpoint(customApiEndpoint);
+      }
+    }
 
-    Appboy.configure(mApplicationContext, configBuilder.build());
+    final boolean enableRequestFocusFix = cordovaPreferences.getBoolean(ENABLE_CORDOVA_WEBVIEW_REQUEST_FOCUS_FIX_PREFERENCE, true);
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.P && enableRequestFocusFix) {
+      // Addresses Cordova bug in https://issuetracker.google.com/issues/36915710
+      BrazeInAppMessageManager.getInstance().setCustomInAppMessageViewWrapperFactory(new CordovaInAppMessageViewWrapper.CordovaInAppMessageViewWrapperFactory());
+    }
+
+    Braze.configure(mApplicationContext, configBuilder.build());
   }
 
   private boolean handleNewsFeedGetters(String action, JSONArray args, final CallbackContext callbackContext) throws JSONException {
     IEventSubscriber<FeedUpdatedEvent> feedUpdatedSubscriber = null;
     boolean requestingFeedUpdateFromCache = false;
 
-    final Appboy mAppboy = Appboy.getInstance(mApplicationContext);
+    final Braze braze = Braze.getInstance(mApplicationContext);
     final String callbackId = callbackContext.getCallbackId();
 
     switch (action) {
@@ -462,7 +501,7 @@ public class AppboyPlugin extends CordovaPlugin {
           }
 
           // Remove this listener from the map and from Appboy
-          mAppboy.removeSingleSubscription(mFeedSubscriberMap.get(callbackId), FeedUpdatedEvent.class);
+          braze.removeSingleSubscription(mFeedSubscriberMap.get(callbackId), FeedUpdatedEvent.class);
           mFeedSubscriberMap.remove(callbackId);
         };
         requestingFeedUpdateFromCache = true;
@@ -477,7 +516,7 @@ public class AppboyPlugin extends CordovaPlugin {
           }
 
           // Remove this listener from the map and from Appboy
-          mAppboy.removeSingleSubscription(mFeedSubscriberMap.get(callbackId), FeedUpdatedEvent.class);
+          braze.removeSingleSubscription(mFeedSubscriberMap.get(callbackId), FeedUpdatedEvent.class);
           mFeedSubscriberMap.remove(callbackId);
         };
         requestingFeedUpdateFromCache = true;
@@ -499,7 +538,7 @@ public class AppboyPlugin extends CordovaPlugin {
           }
 
           // Remove this listener from the map and from Appboy
-          mAppboy.removeSingleSubscription(mFeedSubscriberMap.get(callbackId), FeedUpdatedEvent.class);
+          braze.removeSingleSubscription(mFeedSubscriberMap.get(callbackId), FeedUpdatedEvent.class);
           mFeedSubscriberMap.remove(callbackId);
         };
         requestingFeedUpdateFromCache = false;
@@ -510,12 +549,12 @@ public class AppboyPlugin extends CordovaPlugin {
     if (feedUpdatedSubscriber != null) {
       // Put the subscriber into a map so we can remove it later from future subscriptions
       mFeedSubscriberMap.put(callbackId, feedUpdatedSubscriber);
-      mAppboy.subscribeToFeedUpdates(feedUpdatedSubscriber);
+      braze.subscribeToFeedUpdates(feedUpdatedSubscriber);
 
       if (requestingFeedUpdateFromCache) {
-        mAppboy.requestFeedRefreshFromCache();
+        braze.requestFeedRefreshFromCache();
       } else {
-        mAppboy.requestFeedRefresh();
+        braze.requestFeedRefresh();
       }
     }
 
@@ -527,20 +566,20 @@ public class AppboyPlugin extends CordovaPlugin {
     final IEventSubscriber<ContentCardsUpdatedEvent> subscriber = new IEventSubscriber<ContentCardsUpdatedEvent>() {
       @Override
       public void trigger(ContentCardsUpdatedEvent event) {
-        Appboy.getInstance(mApplicationContext).removeSingleSubscription(this, ContentCardsUpdatedEvent.class);
+        Braze.getInstance(mApplicationContext).removeSingleSubscription(this, ContentCardsUpdatedEvent.class);
 
         // Map the content cards to JSON and return to the client
         callbackContext.success(ContentCardUtils.mapContentCards(event.getAllCards()));
       }
     };
-    Appboy.getInstance(mApplicationContext).subscribeToContentCardsUpdates(subscriber);
+    Braze.getInstance(mApplicationContext).subscribeToContentCardsUpdates(subscriber);
     final boolean updateFromCache = action.equals(GET_CONTENT_CARDS_FROM_CACHE_METHOD);
-    Appboy.getInstance(mApplicationContext).requestContentCardsRefresh(updateFromCache);
+    Braze.getInstance(mApplicationContext).requestContentCardsRefresh(updateFromCache);
     return true;
   }
 
   private boolean handleContentCardsLogMethods(String action, JSONArray args, final CallbackContext callbackContext) {
-    final Appboy appboy = Appboy.getInstance(mApplicationContext);
+    final Braze braze = Braze.getInstance(mApplicationContext);
     final String cardId;
 
     if (args.length() != 1) {
@@ -561,7 +600,7 @@ public class AppboyPlugin extends CordovaPlugin {
     // Only obtaining the current list of cached cards is ok since
     // no id passed in could refer to a card on the server that isn't
     // contained in the list of cached cards
-    final List<Card> cachedContentCards = appboy.getCachedContentCards();
+    final List<Card> cachedContentCards = braze.getCachedContentCards();
 
     // Get the desired card by its id
     final Card desiredCard = ContentCardUtils.getCardById(cachedContentCards, cardId);
@@ -592,7 +631,7 @@ public class AppboyPlugin extends CordovaPlugin {
   private void initializeGeofences() {
     String fineLocationPermission = "android.permission.ACCESS_FINE_LOCATION";
     if (Build.VERSION.SDK_INT >= 29) {
-      String accessBackgroundPermission = "android.permission.ACCESS_FINE_LOCATION";
+      String accessBackgroundPermission = "android.permission.ACCESS_BACKGROUND_LOCATION";
 
       if (cordova.hasPermission(fineLocationPermission) && cordova.hasPermission(accessBackgroundPermission)) {
         AppboyLocationService.requestInitialization(mApplicationContext);
@@ -623,10 +662,10 @@ public class AppboyPlugin extends CordovaPlugin {
   private void requestLocationPermission(final CallbackContext callbackContext) {
     String fineLocationPermission = "android.permission.ACCESS_FINE_LOCATION";
     if (Build.VERSION.SDK_INT >= 29) {
-      String accessBackgroundPermission = "android.permission.ACCESS_FINE_LOCATION";
+      String accessBackgroundPermission = "android.permission.ACCESS_BACKGROUND_LOCATION";
       // Get location permissions, if we need them
       if (cordova.hasPermission(fineLocationPermission) && cordova.hasPermission(accessBackgroundPermission)) {
-        AppboyLocationService.requestInitialization(mApplicationContext);
+        Braze.getInstance(mApplicationContext).requestLocationInitialization();
         callbackContext.success("Location permission already granted. Started location service.");
       } else {
         // Request the permission
@@ -636,14 +675,16 @@ public class AppboyPlugin extends CordovaPlugin {
     } else {
       // Get location permissions, if we need them
       if (cordova.hasPermission(fineLocationPermission)) {
-        AppboyLocationService.requestInitialization(mApplicationContext);
+        Braze.getInstance(mApplicationContext).requestLocationInitialization();
         callbackContext.success("Location permission already granted. Started location service.");
       } else {
         // Request the permission
         locationPermissionCallbackContext = callbackContext;
         cordova.requestPermission(this, LOCATION_REQUEST_CODE, fineLocationPermission);
       }
-    }  
+    }
+    
+    
   }
   
   private void getLocationPermission(final CallbackContext callbackContext) {
@@ -695,37 +736,6 @@ public class AppboyPlugin extends CordovaPlugin {
     return array;
   }
 
-  private static Month parseMonth(int monthInt) {
-    switch (monthInt) {
-      case 1:
-        return Month.JANUARY;
-      case 2:
-        return Month.FEBRUARY;
-      case 3:
-        return Month.MARCH;
-      case 4:
-        return Month.APRIL;
-      case 5:
-        return Month.MAY;
-      case 6:
-        return Month.JUNE;
-      case 7:
-        return Month.JULY;
-      case 8:
-        return Month.AUGUST;
-      case 9:
-        return Month.SEPTEMBER;
-      case 10:
-        return Month.OCTOBER;
-      case 11:
-        return Month.NOVEMBER;
-      case 12:
-        return Month.DECEMBER;
-      default:
-        return null;
-    }
-  }
-
   /**
    * Parses the preference that is optionally prefixed with a constant.
    *
@@ -734,7 +744,7 @@ public class AppboyPlugin extends CordovaPlugin {
   private static String parseNumericPreferenceAsString(String preference) {
     if (preference != null && preference.startsWith(NUMERIC_PREFERENCE_PREFIX)) {
       String preferenceValue = preference.substring(NUMERIC_PREFERENCE_PREFIX.length(), preference.length());
-      AppboyLogger.d(TAG, "Parsed numeric preference " + preference + " into value: " + preferenceValue);
+      BrazeLogger.d(TAG, "Parsed numeric preference " + preference + " into value: " + preferenceValue);
       return preferenceValue;
     }
     return preference;
@@ -750,7 +760,7 @@ public class AppboyPlugin extends CordovaPlugin {
 
     if (preference != null && preference.startsWith(NUMERIC_PREFERENCE_PREFIX)) {
       preferenceValue = preference.substring(NUMERIC_PREFERENCE_PREFIX.length(), preference.length());
-      AppboyLogger.d(TAG, "Parsed numeric preference " + preference + " into value: " + preferenceValue);
+      BrazeLogger.d(TAG, "Parsed numeric preference " + preference + " into value: " + preferenceValue);
     }
 
     // Parse the string as an integer. Note that this is the same decoding used in CordovaPreferences