+++ src/ios/AppboyPlugin.m
@@ -65,6 +65,7 @@
             inApplication:notification.object
         withLaunchOptions:notification.userInfo
         withAppboyOptions:appboyLaunchOptions];
+  [[Appboy sharedInstance] addSdkMetadata:@[ABKSdkMetadataCordova]];
 
   if (![self.disableAutomaticPushRegistration isEqualToString:@"YES"]) {
     UIUserNotificationType notificationSettingTypes = (UIUserNotificationTypeBadge | UIUserNotificationTypeAlert | UIUserNotificationTypeSound);
@@ -129,7 +130,7 @@
 }
 
 - (void)requestImmediateDataFlush:(CDVInvokedUrlCommand *)command {
-  [[Appboy sharedInstance] flushDataAndProcessRequestQueue];
+  [[Appboy sharedInstance] requestImmediateDataFlush];
 }
 
 - (void)requestLocationPermission:(CDVInvokedUrlCommand *)command {
@@ -218,10 +219,18 @@
 
 - (void) setGender:(CDVInvokedUrlCommand *)command{
   NSString *gender = [command argumentAtIndex:0 withDefault:nil];
-  if ([gender.lowercaseString isEqualToString:@"m"]) {
-    [[Appboy sharedInstance].user setGender:ABKUserGenderMale];
-  } else if ([gender.lowercaseString isEqualToString:@"f"]) {
+  if ([gender.lowercaseString isEqualToString:@"f"]) {
     [[Appboy sharedInstance].user setGender:ABKUserGenderFemale];
+  } else if ([gender.lowercaseString isEqualToString:@"m"]) {
+    [[Appboy sharedInstance].user setGender:ABKUserGenderMale];
+  } else if ([gender.lowercaseString isEqualToString:@"n"]) {
+    [[Appboy sharedInstance].user setGender:ABKUserGenderNotApplicable];
+  } else if ([gender.lowercaseString isEqualToString:@"o"]) {
+    [[Appboy sharedInstance].user setGender:ABKUserGenderOther];
+  } else if ([gender.lowercaseString isEqualToString:@"p"]) {
+    [[Appboy sharedInstance].user setGender:ABKUserGenderPreferNotToSay];
+  } else if ([gender.lowercaseString isEqualToString:@"u"]) {
+    [[Appboy sharedInstance].user setGender:ABKUserGenderUnknown];
   }
 }
 
@@ -382,6 +391,20 @@
   NSString *language = [command argumentAtIndex:0 withDefault:nil];
   if (language != nil) {
     [Appboy sharedInstance].user.language = language;
+  }
+}
+
+- (void) addToSubscriptionGroup:(CDVInvokedUrlCommand *)command {
+  NSString *groupId = [command argumentAtIndex:0 withDefault:nil];
+  if (groupId != nil) {
+    [[Appboy sharedInstance].user addToSubscriptionGroupWithGroupId:groupId];
+  }
+}
+
+- (void) removeFromSubscriptionGroup:(CDVInvokedUrlCommand *)command {
+  NSString *groupId = [command argumentAtIndex:0 withDefault:nil];
+  if (groupId != nil) {
+    [[Appboy sharedInstance].user removeFromSubscriptionGroupWithGroupId:groupId];
   }
 }